* reworked xtc component:
- expose only a pre and post filter;
- expose frequency, type, and Q for each filter;
- expose a preGain and postGain which change filter gain if type uses it, else change a _preGain or _postGain node in the component's connection graph;
- consider making these filters AudioComponents in their own right, thereby getting bypass functionality for free; currently there is no way to bypass these pre and post filters

** need to consider how parameters are being handled

currently all components which have a UI must expose all parameters as numbers, strings, or booleans. could be expose AudioParams as well? this would relieve the need to write getters and setters which operate on the value of AudioParams directly; we could expose the AudioParam via the getter and wouldn't need the setter. this could also open the way for better automation...